# Improve merging of docs on generation