Cleans up and improves (slightly, it seems) performance in merging docs by using sorted lookup

---

* Closes: https://github.com/godotengine/godot-proposals/issues/9125

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->