@nlupugla 2024-02-18
Since the merging logic is non-trivial now, I might suggesting making a helper function or iterator that you can reuse for the different merge cases. For example, a `merge_vectors` method that takes a function pointer, which you might call like `merge_vectors(to_properties, from_properties, &merge_property)`, where `merge_property` is a function that just handles merging two matched properties.

If you feel that overcomplicates things, I think repeating the iteration code is fine as well.