@AThousandShips 2024-02-18
The comparison isn't the same in the different cases, so it'd be more complicated

Also realized I need to fix some details of this, one moment