@AThousandShips 2024-02-18
Fixed some of the code, getting late so will continue working tomorrow, should work now, but with some new changes, will elaborate on them tomorrow

Updated the benchmarks above