@nlupugla 2024-02-18
Wow, the new benchmarks show a bigger improvement and it seems like most of it is due to the `saving` somehow getting faster. It doesn't look like any of your code directly affects saving, unless somehow ensuring the docs are sorted makes saving faster. Maybe you your disk drive was less busy during the later tests or something?